Use bot_flux_to_tend from tracer_vertdiff
This commit still uses the KMT kludge, but now computes bot_flux_to_tend(:) by
applying a unit bottom flux to a column of 0s in tracer_vertdiff. This commit
includes a check to ensure that sum(dz(:) * bot_flux_to_tend(:)) == 1, though I
am not sure if we want that in the code base long-term.

This commit also fixes the units of the sfc_flux argument passed to
tracer_vertdiff() -- it should be Rho0 * STF instead of just STF. This requires
an update to MOM_tracer_diabatic (merged to dev/NCAR in 276954f but not yet on
this branch); for testing, I copied the updated version of that file but am not
committing it to make the future merge easier.

Last note: I believe the btm_flux argument to tracer_vertdiff is actually
positive upward, although the comments in MOM_tracer_diabatic.F90 claim it is
negative upward. I'll investigate a little more and open an issue ticket with
GFDL if that turns out to be the case, but in this commit I'm multiplying
bot_flux_to_tend(:) by -m_per_cm instead of just m_per_cm and I suspect it's
due to the sign of btm_flux