Only define KPP tracer diags when using KPP
The three per-tracer diagnostics added to the tracer registry (surface flux,
non-local term, and budget component of NLT) are only defined if KPP is enabled
in the diabatic driver. Note that this requires a change to MOM.F90, which may
not be desired -- the alternative is to always register the KPP diagnostics
whether KPP is enabled or not.