Can pass tr_ptr to KPP_NonLocalTransport()
Added tr_ptr to the thermo_var_ptrs type, and updated MOM.F90 to have them
point to T & S correctly. There is some kludgy logic in
register_tracer_diagnostics() to register the T & S diagnostics correctly, but
I'll clean that up in a future commit. In the mean time, I was able to move
some of the diagnostic IDs out of the KPP control structure.