Major refactor of MARBL forcing fields
Removed marbl_forcing_type and marbl_ice_ocean_boundary_type, keeping the flat
structure of forcing and ice_ocean_boundary_type, respectively. This greatly
reduces the amount of code in marbl_forcing_type_mod.F90 (which now needs a new
name), and adds a little more overhead to the driver layer but with the benefit
of reusing pointers for things like ice_fraction and u10_sqr which are needed
in the CFC cap as well.

I also started the process of adding multiple ice category support, but there's
more to do for that feature.