Update IC file and units of RIV_FLUXES
The old MARBL initial condition file had some issues at depth when the MOM
topography was deeper than POP; these cells were all set to 0 because the
restart file we mapped tracers from does not include a land mask. Fix was to
mask out the POP data prior to mapping it, then the lateral fill was applied
correctly.

Additionally, Keith noticed a bug in our call to
applyTracerBoundaryFluxesInOut() -- the in_flux_optional argument should be in
units of conc m, but we were passing CS%RIV_FLUXES with units of conc m/s.
Since the in_flux should be the time-integrated value, we set

CS%RIV_FLUXES = CS%RIV_FLUXES * dt

prior to passing it to applyTracerBoundaryFluxesInOut()

The comments in the declaration of RIV_FLUXES was also updated to account for
the unit change.