Add MARBL_TRACERS_INIT_VERTICAL_REMAP_ONLY option
If the initial condition file for the MARBL tracers is already on the MOM grid,
this will skip the horizontal interpolation step. This option is not necessary
for CESM-MOM6 (we want to interpolate ICs from the WOA grid), but is useful for
MOM6-examples, where we are interpolating to a single column grid offline.