Use time_interp_external for restoring
This is still a work in progress, but I cleaned up how the MARBL tracer
restoring fields are translated into something the MARBL_tracers control
structure can parse and also call init_external_field. One issue I'm having is
that time_interp_external needs a time_type argument, and that's not available
in MARBL_tracers_column_physics. I think the solution will be to move a lot of
this code to MARBL_forcing.F90, but I want to do that in a separate commit.