move post_data calls for river nutrient fluxes
Diags are posted after files are read in set_forcing(). Also, cleaned up where
I apply the dt factor to convert from flux to time-integrated flux because
set_forcing is only called once during the first two time steps so we were
inadvertently applying a factor of dt^2 on the second time step.

This commit is bit-for-bit with 69202f7 in my CESM testing, which wasn't the
case for 5540525