Check abio_dic_on and base_bio_on before posting
There are a few diagnostics that are only defined if base_bio_on=.true. (the
river flux nutrient forcing fields), and one that is only defined if
abio_dic_on=.true. (the d14c forcing); some compilers won't initialize the
diagnostic ids to 0 in the control structure, so we need to either explicitly
initialize all the ids or only call post_data when we know the ids have been
set. This commit does the latter.