Use MARBL's autotroph_tracer_consistency_enforce()
There was an issue where smoothing the MARBL tracer initial conditions led to
some inconsistent tracer values for tracers associated with MARBL's autotrophs.
For each autotroph, if any tracer (C, N, P, etc) is 0 we want all of that
autotroph's tracers to be 0... but the smoothing algorithm in
MOM_initialize_tracer_from_Z() would sometimes leave one tracer at 0 but
introduce a nonzero value in a different tracer at the same location. The
subroutine on the MARBL interface enforces this consistency.

This function is called whenever any MARBL tracer is initialized with
MOM_initialize_tracer_from_Z(); if all the tracers are read in from a restart
file then this subroutine is skipped. If this function is called, it is noted
in the log with the following message:

Enforcing consistency across autotroph tracer initial conditions