Use prediabatic SST and SSS for MARBL surface flux
The existing MARBL driver uses T & S from the end of the diabatic routine as
inputs into surface_flux_compute() and interior_tendency_compute(). For
surface_flux_compute(), there is a discrepancy because we have not yet called
tracer_vertdiff for the MARBL tracers but have for T & S. To enforce
consistency among all the tracers, we want to use the pre-vertdiff T&S
quantities.

There is a little bit of infrastructure change -- I added 3D fields for T & S
at the beginning of the diabatic subroutine to the diabatic control structure
(we will eventually want to call interior_tendency_compute() before
tracer_vertdiff as well), and then had to add a routine to
MOM_tracer_flow_control.F90 to extract use_MARBL_tracers from the tracer flow
control structure. prediabatic_T and prediabatic_S are now optional arguments
to call_tracer_column_fns(), but that function will abort if USE_MARBL_TRACERS
is true and the arguments are not present because they are required in
MARBL_tracers_column_physics()